Fixing warnings in uts benchmark; see #114

+++ omp-tasks/uts/brg_sha1.c
@@ -60,10 +60,10 @@ void rng_init(RNG_state *newstate, int seed)
 
   for (i=0; i < 16; i++) 
     gen.state[i] = 0;
-  gen.state[16] = 0xFF & (seed >> 24);
-  gen.state[17] = 0xFF & (seed >> 16);
-  gen.state[18] = 0xFF & (seed >> 8);
-  gen.state[19] = 0xFF & (seed >> 0);
+  gen.state[16] = (u_int8_t) (0xFF & (seed >> 24));
+  gen.state[17] = (u_int8_t) (0xFF & (seed >> 16));
+  gen.state[18] = (u_int8_t) (0xFF & (seed >> 8));
+  gen.state[19] = (u_int8_t) (0xFF & (seed >> 0));
   
   sha1_begin(&ctx);
   sha1_hash(gen.state, 20, &ctx);
@@ -73,12 +73,12 @@ void rng_init(RNG_state *newstate, int seed)
 void rng_spawn(RNG_state *mystate, RNG_state *newstate, int spawnnumber)
 {
 	struct sha1_context ctx;
-	uint8  bytes[4];
+	u_int8_t  bytes[4];
 	
-	bytes[0] = 0xFF & (spawnnumber >> 24);
-	bytes[1] = 0xFF & (spawnnumber >> 16);
-	bytes[2] = 0xFF & (spawnnumber >> 8);
-	bytes[3] = 0xFF & spawnnumber;
+	bytes[0] = (u_int8_t) (0xFF & (spawnnumber >> 24));
+	bytes[1] = (u_int8_t) (0xFF & (spawnnumber >> 16));
+	bytes[2] = (u_int8_t) (0xFF & (spawnnumber >> 8));
+	bytes[3] = (u_int8_t) (0xFF & spawnnumber);
 	
 	sha1_begin(&ctx);
 	sha1_hash(mystate, 20, &ctx);
